리펙토링 전 rdb 장바구니 버전

+++ src/main/java/com/nhnacademy/back/cart/service/CartService.java
@@ -11,7 +11,7 @@ import com.nhnacademy.back.cart.domain.dto.response.ResponseCartItemsForMemberDT
 
 public interface CartService {
 	void createCartForMember(String memberId);
-	int createCartItemForMember(RequestAddCartItemsDTO request);
+	Long createCartItemForMember(RequestAddCartItemsDTO request);
 	int updateCartItemForMember(long cartItemId, RequestUpdateCartItemsDTO request);
 	void deleteCartItemForMember(long cartItemId);
 	void deleteCartForMember(String memberId);
